backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.117.1",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "ffmpeg-python>=0.2.0",
]
